\name{MultiFlow-package}
\alias{MultiFlow-package}
\alias{MultiFlow}
\docType{package}
\title{
Multiplex Lateral Flow Assays.
}
\description{
Image analysis (cropping, segmetation, thresholding, feature engineering) of 
images taken from lateral flow assays, computation of linear calibration models, 
automatic report generation via rmarkdown by means of a shiny app.
}
\details{
\tabular{ll}{
Package: \tab MultiFlow\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2020-08-30\cr
Depends: \tab R(>= 4.0.0)\cr
Imports: \tab stats, shiny, shinyjs, shinythemes, shinyFiles, EBImage, DT, rmarkdown, ggplot2\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MultiFlow)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
