\name{runMultiFlowApp}
\alias{runMultiFlowApp}
\title{ Run MultiFlow Shiny App }
\description{
  The function startes the MultiFlow shiny app.
}
\usage{
runMultiFlowApp()
}
%\arguments{}
\details{
  The function startes the MultiFlow shiny app. 
}
\value{ No return value }
\references{ 
  C. Ruppert, L. Kaiser, L.J. Jacob, S. Laufer, M. Kohl, H.P. Deigner (2020).
  Duplex shiny app quantification of sepsis biomarkers C-reactive protein and 
  interleukin-6 in a fast quantum dot labeled lateral flow assay. In revision.
}
\author{ Matthias Kohl \email{Matthias.Kohl@stamats.de}}
\note{ The app was implemented based on the app provided by package ShinyImage 
       and a modification of the ShinyImage app implemented by Weronika Schary 
       and Filip Paskali. }
\examples{
\dontrun{
   runMultiFlowApp()
}
}
\keyword{dynamic}
\keyword{datagen}
\concept{shiny app}
